Add all-solutions capability, add options, add test cases
The solver can now optionally find all distinct solutions to a numbers
game. This is slower than the default fast mode, firstly because it
doesn't stop when it finds the first solution like fast mode does,
and secondly because it isn't so aggressive with removing redundant
expressions.

For the fast mode, the expressions (10-3)*5 and 3*10+5 are considered
equivalent, because they use the same starting numbers and reach the
same total, so it knows it doesn't need both. But the all-solutions mode
considers them to be separate expressions, so they're both used in the
search. Two solutions which differ by one using (10-3)*5 and the other
using 3*10+5 are considered different solutions.

The all-solutions mode ensures every expression has its components
ordered in a standard way, so that trivial rearrangements of the same
expression don't occur (e.g. 3*4+5 and 5+4*3).

There's an options button in the top right. Currently it only allows
the user to switch between fast mode and all-solutions mode, and read
the About page.

Even if you run a solve in fast mode, a link will appear in the output
pane offering to search for more solutions. If the user clicks that link
it will run the solver in all-solutions mode on the same selection and
target.

Finally, there's test.html, which uses solver_tests.js. This runs the
solver against a set of 250 test cases, which were randomly generated
and checked with another solver.