Pass single objects to callbacks rather than individual arguments
The progressCallback and finishedCallback which the interface passes to
the engine when starting a solve both take a list of arguments. This
makes it hard to add more information to the progress callback or the
solve result in a backwardly-compatible way.

So before it becomes a problem, let's change the API so that
progressCallback() is expected to take one argument (a SolverProgress
object), and finishedCallback() is expected to take one argument
(a SolverResults object).

The SolverProgress and SolverResults objects can have public APIs to get
the actual information out of them, and we can easily add new methods to
these objects in the future without breaking any existing
implementations.

The SolverProgress object now contains something that wasn't passed back
to the application before - the number of "best" solutions it has found
so far.