Limit selections to 7 numbers in all-solutions mode
If you run a solve on an 8-number selection in all-solutions mode,
we run out of memory part way through generating the 7-number
expressions.

So first, if we're in all-solutions (full) mode, limit the selection to
7 numbers.

Second, if we've reached the stage where we're generating expressions
with N numbers, where N is the length of the selection, currently we
remember all those expressions, just like we remember all the shorter
expressions.

However, we don't actually have to remember an N-number expression
unless it equals the target or is at least as close as our closest
expression so far. An N-number expression which isn't a candidate
solution is useless to us - we're never going to need to combine
N-number expressions with other expressions because we'll never need
to generate expressions longer than N numbers.

This means we can run in all-solutions mode if we have 7 numbers,
because we now throw away most of the 7-number expressions we
generate.

Finally, use arrays in a few places where we're currently using objects,
where the keys are only ever going to be integers.