Optimise expression building loop, save half the work
The solver works by starting with the 6 starting numbers (we'll stick
with 6 for this example), then building all possible 2-number
expressions from that, then building all possible 3-number expressions
from the 2-number and 1-number expressions it's already got, and so on.

When it comes to build the 6-number expressions, first it'll combine all
the 1-number expressions with the 5-number expressions, then all the 2s
with the 4s, then all the 3s with all the 3s, then all the 4s with all
the 2s, then all the 5s with all the 1s.

In fact we could have stopped once we've combined all the 3s with the
3s. After that, we're just generating the same pairs we've already
generated, but with the expressions in each pair the other way round. We
never need to generate the same pair of expressions both ways round - in
fact we can just ensure we put the higher-valued expression on the left
of the operator every time. This will have no effect on addition or
multiplication, and for subtraction and division we always have to put
them that way round anyway - Countdown doesn't allow negative or
non-integer numbers at any stage of the calculation.

The redundant pairs we currently generate do of course get discarded
one way round or the other because we've already got an equivalent
expression, so this won't save any memory, but it will save a lot of
time by not having to generate them in the first place.

So if we're generating the N-number expressions, we only need to
generate pairs of expressions of length (1, N-1), (2, N-2), ... up to
(N/2, N/2). If N is odd, we go up to ((N-1)/2, (N+1)/2).