Uselessness analysis to remove things like 100 + 10 - 5
Subject every OrderedExpression to an isUseless() check. If an
expression is considered "useless", it is discarded for all purposes,
even the mode that searches for all solutions.

A division expression, consisting of the product of one or more
factor expressions divided by the product of one or more factor
expressions, is useless if either of the following apply:
    * Any factor expression on the numerator is the square of the
      value of the denominator. For example, (100 * 25) / (4 + 1)
      is useless because 25 is the square of (4 + 1), and this is
      trivially simplified to 100 * (4 + 1).
    * There is any subset of factor expressions on the numerator whose
      product equals the value of the whole expression. For example,
      (100 * 3 * 4 * 7) / (7 + 5) = 700 is useless because 100 * 7
      equals 700.

A subtraction expression, consisting of the sum of one or more
terms (the left hand side, or LHS) minus the sum of one or more terms
(the right hand side, or RHS), is useless of either of the following
apply:
    * Any term on the LHS is twice the value of the RHS. For example,
      (100 + 10) - 5 is useless because 10 is twice 5, so the expression
      could be simplified to 100 + 5.
    * There is any subset of terms on the LHS whose sum equals the
      value of the whole expression. For example,
      (100 + 7 + 2) - (3 + 4) = 102 is useless because 100 + 2 = 102.

This fixes cases where the solver generated solutions involving
subtracting 5 from 10, for example. There was an existing check which
would prohibit generating a minus operation between two expressions
where the LHS was twice the RHS, but that didn't catch cases where
the LHS was (100 + 10) and the RHS was 5 or 50, for example.

BinaryExpression objects are not subject to uselessness analysis,
because if we're using those then we're in fast mode, and that
terminates as soon as a solution is found, which will always be a
shortest solution anyway.