Ability to solve all targets at once, and improve all-sols algorithm
The solverRunAllSolutions() function can now take a range of other
targets (imperfectSolutionsMin and imperfectSolutionsMax) which causes
the solver to return more than just the nearest solutions to the target.

All expressions generated whose value falls within that range are saved
in the SolverResults object, and this can be obtained by the
SolverResults object's getImperfectSolutions() method. This returns a
map of targets to lists of Expressions.

This means the target passed to solverRunAllSolutions() can now be null,
meaning there's no target and SolverResults.getSolutions() will return
an empty array. However, expressions whose value fall in the range
[imperfectSolutionsMin, imperfectSolutionsMax] can be obtained by
SolverResults.getImperfectSolutions().

The algorithm for finding all solutions has been improved so that we
don't need a map of counts of numbers used any more. This used a lot
of memory and reduced the performance. Instead, for each expression we
remember a list of bitmasks, where each bitmask represents a possible
set of starting numbers that could be used to make this expression.
To determine whether two expressions can be combined, we check there is
a bitmask from one which has no bits in common with at least one bitmask
from the other.

There is also a new function solverRunAllTargets(). This uses the fast
binary-tree algorithm (so not the all-solutions algorithm, and hence
only one solution will be returned per target), but continues until
we've generated all expressions. This can be used in conjunction with
the imperfectSolutionsMin and imperfectSolutionsMax parameters to
cheaply generate a list of all targets achievable with a given
selection, with one example solution given for each possible target.

Internally, the three strategies are called STRATEGY_FAST_CUT (original
fast algorithm: we don't attempt to find all the solutions, and we
finish early if we find an expression that matches the target),
STRATEGY_FAST (fast algorithm, but we continue until we've checked every
expression in the solution space, so we know for every target whether
it's possible or not with this selection), and STRATEGY_ALL_SOLUTIONS
(improved all-solutions algorithm, finds all non-trivially different
solutions for each target).