Pattern-match solution trees to categorise them into solution type
Introduce one user-visible function, getCategoryName(), which takes
an expression (e.g. (4 + 1) * 100 + 6 * 7) and returns a one-line
human-readable description of what general type of solution it is
(in this case, "500 + 42"). It could be more complex, for example,
the category for (25*75+1)*50/100 is "4L(937.5+0.5)".

This function hides a complex and messy system of pattern matching where
we try to match the expression tree against various patterns, from the
specific to the more general, and return a description string for the
first one that works.

Each pattern extends SolutionPattern and defines a match() function
which should return true if the given expression matches the pattern. If
the expression matches the pattern, it may also set arbitrary
(name -> value) pairs in a map (expNames), the meanings (and value types)
of which are pattern-specific.

Every pattern must, on a match, call SolutionPattern.acceptMatch() which
sets expNames[name] to the expression, where name is the name given to
the pattern when it was constructed.