Fix duplicate solution bug
To reproduce:: 8 7 2 10 10 6 -> 982, finding all solutions.

This gives the solution "(8 + 6) × 10 × 7 + 2" twice.

The reason is because each solution uses a different 10.

We store a list of bitmasks (representing starting numbers used) for
each expression we generate. Often each expression has only one bitmask,
but if a number is repeated, an expression might have two or more,
because it could use any one (or any subset) of the repeated numbers.
For example, if there are two 5s, the expression (3 + 5) is compatible
with (7 * 5) because the expressions could use different 5s, so the
array of bitmasks for each expression need to reflect that so we can
see the expressions are in fact compatible.

However, the single number expressions still use a single bitmask,
and they get added to the expression list once for each time the
number appears in the selection. This is causing us to generate
duplicate solutions - two solutions whose string representation is
the same but which use "different" starting numbers.

So the SingleNumber expressions now take a list of bitmasks, and if
there are two 10s in the selection, we add to the expression list one
SingleNumber object with two one-bit bitmasks, indicating that this
SingleNumber expression could take either of those two bitmasks (i.e.
use either of the 10s).