Introduce desktop browser version of numbers solver interface
This interface can solve all targets for a solution simultaneously and
display them in a map. If you ask it to display all the solutions for a
specific target, it will do so and categorise the solutions using the
solution categoriser introduced a few commits ago.

This new interface uses the same engine as the original solver interface
which was intended for mobile devices, but this one is a bit more
featureful.