Allow linking to specific solution pages
The script for the desktop solver now looks at the URL's query string to
see if there's a puzzle in it. Parameters are:

sel: the numbers in the selection, separated by hyphens.
target: the target, if specified.
fast: if set, use the fast solver.

Each solution report or target map now has a "link to this page" link
which can be copied and shared.