Reorganise files into directories under resources/
Create directories resources/css, resources/js and resources/images into
which we've moved the non-HTML files as appropriate. This makes it a bit
easier to have both the mobile and desktop solver on a site, with the
mobile and desktop HTML pages in different directories each having a
resources symlink pointing to one place. This means when we update the
JavaScript, CSS or images, we can just upload them to one place and both
solver interfaces will pick up the changes.