Support for other numbers variants using min, max and lock
The selection box now supports additional parameters, min and max, to
specify that a valid solution must use at least or at most a certain
number of starting numbers. These are specified as, say, min=3 or max=5
in the selection box or in the query string. This opens the door to
solving Omelette games ("min=6", so you must use all the numbers) and
Duckdown games ("max=6" with a seven-number selection, meaning you must
not use all the numbers).

Any number in the selection may now be preceded with an "L" to indicate
that it is a "locked" number in a Lockdown variant, which means that
number must be used for a solution to be valid. In the query string
this feature looks a bit different - it is specified as a
hyphen-separated list of locked numbers as the value to the parameter
"lock".

For example, in the selection box, you might see:
100 1 1 8 6 L9

In the query string, it would look like this:
...?sel=100-1-1-8-6-9&target=557&lock=9

These are explained in the help window.

If a minimum number of numbers ("min") or any locked numbers are
specified, the solver will no longer ignore apparently useless
operations like multiplying or dividing by 1, because these operations
might not be useless any more. This means that (a) you can't use "min"
or "lock" with the fast solver, and (b) the number of solutions returned
might be unexpectedly greater if you use either of these constraints.

This also has the effect that Hyperlock (8 numbers, of which one is
locked) isn't supported, because the fast solver doesn't support locked
numbers and the fast solver is the only one currently supported for
eight-number selections.

In solver_engine.js, the solverRun() function now has an additional
optional argument, maxNumbersUsed, which defaults to null.
solverRunAllTargets() and solverRunAllSolutions() have three additional
optional arguments: minNumbersUsed (default null), maxNumbersUsed
(default null) and lockedNumbers (an array of the locked numbers,
default []).

Defaults and limits:

min: default 1, must be at least 1.
max: default is the number of numbers in the selection (i.e. no
restriction), must be at least the value of min.
lock: default none, must be a subset of the numbers in the selection.