Use fast&cut strategy by default when there are 8 numbers
With an 8-number selection, we use the fast solver but we still
generate solutions for every possible target, on the off-chance
that the user is going to scroll to it using the ribbon displayed
either side of the target. With 8 numbers this can take a
considerably long time, especially if there are lots of
divisions available. It also uses a huge amount of memory, more
than might reasonably be available.

So for an 8-number selection with a specified target we now default
to the FAST_CUT strategy, which not only uses the fast solver but
also quits when it's found an exact solution to the specified target.
This means there's no target ribbon when 8 numbers are given.

If you want the old behaviour, add "all" to the selection box, and
it will generate solutions to all targets as before. This also
happens if you click Map.

"cut" may also be specified in the selection box to force the use of the
FAST_CUT strategy for selections containing fewer than 8 numbers.