Introduce preferences window
In the desktop interface, there is now a button on the right of the
top bar with a gear on it. This opens the preferences window.

There is currently only one preference you can change - whether you
want solutions written out as one line with brackets if necessary,
like "(100 + 7) * 5", or with each individual operation on a separate
line, like "100 + 7 = 107\n107 * 5 = 535". I've borrowed some chess
terms and called these algebraic notation and descriptive notation.

The preferences window also has a toggle option (default off) to say
whether you want to save your preferences in a cookie. If you enable
this (the positive action required to consent to storage of cookies),
preferences are saved to cookies when the preferences window is
closed. If you later disable it, preference cookies are deleted.

Cookie names:
quantum_tombola_cookies: if present and the value is "yes", the user
has consented to preference cookies.
quantum_tombola_notation: "algebraic" or "descriptive". Default is
"algebraic".